Address PR feedback: merged all dashboard page componenets into DashboardHero, reverted Hero.tsx componenet to orginal state, and removed all login content (page, buttons, styling)
