+++ src/App.tsx
@@ -1,6 +1,5 @@
 import { BrowserRouter, Routes, Route, Link, Navigate } from "react-router-dom";
 import Landing from "./pages/Landing";
-import Login from "./pages/Login";
 import Dashboard from "./pages/Dashboard";
 
 function StubPage({ title, description }: { title: string; description: string }) {
@@ -30,7 +29,6 @@ export default function App() {
     <BrowserRouter>
       <Routes>
         <Route path="/" element={<Landing />} />
-        <Route path="/login" element={<Login />} />
         <Route path="/dashboard" element={<Dashboard />} />
         <Route
           path="/schedule"
