+++ src/components/Hero.tsx
@@ -1,11 +1,68 @@
 "use client"
 
 import { siteConfig } from "../config/site"
+import { useState } from "react"
 
 export default function Hero() {
+  const [isDropdownOpen, setIsDropdownOpen] = useState(false)
 
   return (
     <section className="relative min-h-screen w-full flex items-center justify-center overflow-hidden hero-pixel-mountains">
+
+      {/* Navigation with branding */}
+      <nav className="absolute top-6 left-6 right-6 z-20 flex justify-between items-center px-4 py-3 rounded-xl bg-black/40 backdrop-blur-md border border-white/20 shadow-lg">
+        {/* Logo/Brand */}
+        <div className="text-white font-pixel text-lg md:text-xl drop-shadow-lg">
+          <div>
+            <span className="text-valley-gold">TechHack</span> <span className="text-sky-blue">Valley</span>
+          </div>
+          <div className="text-xs text-white/60 font-pixel">
+            by GT Esports
+          </div>
+        </div>
+        
+        {/* Menu Button */}
+        <div className="relative">
+          <button
+            onClick={() => setIsDropdownOpen(!isDropdownOpen)}
+            className="btn-pixel flex items-center gap-2"
+          >
+            Menu
+            <svg
+              className={`w-4 h-4 transition-transform ${isDropdownOpen ? "rotate-180" : ""}`}
+              fill="none"
+              stroke="currentColor"
+              viewBox="0 0 24 24"
+            >
+              <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M19 9l-7 7-7-7" />
+            </svg>
+          </button>
+
+          {isDropdownOpen && (
+            <div className="absolute right-0 mt-2 w-56 card-pixel overflow-hidden z-30 shadow-2xl">
+              <a
+                href="/schedule"
+                className="block px-4 py-3 text-valley-brown hover:bg-valley-gold/20 transition-colors font-pixel text-sm"
+              >
+                📅 Schedule
+              </a>
+              <a
+                href="/sponsors"
+                className="block px-4 py-3 text-valley-brown hover:bg-valley-gold/20 transition-colors font-pixel text-sm"
+              >
+                🏢 Sponsors
+              </a>
+              <a href="/teams" className="block px-4 py-3 text-valley-brown hover:bg-valley-gold/20 transition-colors font-pixel text-sm">
+                👥 Teams
+              </a>
+              <a href="/submit" className="block px-4 py-3 text-valley-brown hover:bg-valley-gold/20 transition-colors font-pixel text-sm">
+                🚀 Submit
+              </a>
+            </div>
+          )}
+        </div>
+      </nav>
+
       {/* Starry background */}
       <div className="absolute inset-0">
         {[...Array(80)].map((_, i) => (
