+++ src/pages/Landing.tsx
@@ -6,12 +6,10 @@ import SponsorsTeaser from "../components/SponsorsTeaser"
 import Community from "../components/Community"
 import FAQ from "../components/FAQ"
 import Footer from "../components/Footer"
-import Navbar from "../components/Navbar"
 
 export default function Landing() {
   return (
     <div className="min-h-screen bg-white w-full">
-      <Navbar/>
       <Hero />
       <EventStats />
       <About />
